+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Switch to dune.
 - Avoid overlapping of boxes (#1).
 - Allow spaces with labels.
+- Ensure correct bounding box for nullary operations.
 
 0.1.0 (2021-01-12)
 =====